+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [CSE] Added validation of &lt;flexContainer>'s *containerDefinition* attribute during CREATE requests.
 - [CSE] Added check for correct Service Provider ID in absolute requests.
 - [CSE] Added support for &lt;pollingChannel>'s *requestAggregation* attribute and functionality.
+- [CSE] Added support for BLOCKING UPDATE notification event type.
 - [WEB] Allow to open the WebUI of a registered CSE via the context menu.
 - [CONSOLE] Added config for dark (default) and light theme for better readability on consoles with light background.
 - [CONSOLE] Added graph plotting for &lt;contentInstance> resources that contain numerical values.